data: Move file review action refusals
File-review action scope resolves command targets, while live-vs-batch and
partial-review refusals build user-facing guidance.

Those refusal paths carried selected-change checks, page formatting, and
command suggestions inside the scope module. That made the scope module
depend on presentation-heavy refusal policy.

This commit addresses that by moving the refusal helpers into
action_refusals and routing command modules plus action scope through the
new module.

Subsequent architecture coverage will record the refusal boundary.
